refactored TelegramService
